[Toolchain] Ensure we don't rebuild kernel headers

+++ Factory/Toolchain/BuildKernelHeaders.sh
@@ -1,11 +1,14 @@
 
 
 pushd $SOURCES_DIR/$KERNEL_HEADERS_NAME
-    make mrproper
-    ARCH=${CONFIG_ARCHITECTURE} make headers
+    if [ ! -f ".built_$PHASE" ]; then
+        make mrproper
+        ARCH=${CONFIG_ARCHITECTURE} make headers
 
-    cp -rv usr/include/* $TOOLS_LOCATION/include
+        cp -rv usr/include/* $TOOLS_LOCATION/include
 
-    find $TOOLS_LOCATION/include \( -name '.*' -o -name '.*.cmd' \) -exec rm -vf {} \;
-    rm -v $TOOLS_LOCATION/include/Makefile
+        find $TOOLS_LOCATION/include \( -name '.*' -o -name '.*.cmd' \) -exec rm -vf {} \;
+        rm -v $TOOLS_LOCATION/include/Makefile
+        touch .built_$PHASE
+    fi
 popd
